refactor: Refactor app structure and modularize code
Moved static, template, and JS/CSS/image files into the app/ directory for better organization. Split main.py into modular files: __init__.py for app setup, database.py for models, routes.py for endpoints, and utils.py for helpers. Updated local-booru.spec to match new paths. This improves maintainability and prepares the project for future expansion.
