+++ app/static/js/batch_actions.js
