+++ app/static/js/gallery.js
