+++ app/static/js/index.js
