+++ app/static/js/notifications.js
