+++ app/static/js/saved_searches_manager.js
