+++ app/static/js/settings.js
