+++ app/static/js/shared_gallery.js
