+++ app/static/js/ui_helpers.js
