+++ app/templates/batch_actions.html
