+++ app/templates/gallery.html
